release 5.15.6-2
as of now i'll be adopting a new versioning scheme for tildearrow
fixes. any new version of this patchset will bump the release
number, e.g. 5.16.0-2. this is to prevent releasing unofficial
versions (like this one) because some of the builds compile the
entire Plasma rather than just this package and I don't want to
break their builds anymore.